requirements.txt
annotated-types==0.6.0 ; python_version >= "3.11" and python_version < "4.0"
anyio==4.2.0 ; python_version >= "3.11" and python_version < "4.0"
attrs==23.2.0 ; python_version >= "3.11" and python_version < "4.0"
beautifulsoup4==4.12.2 ; python_version >= "3.11" and python_version < "4.0"
black==23.12.1 ; python_version >= "3.11" and python_version < "4.0"
boto3==1.34.11 ; python_version >= "3.11" and python_version < "4.0"
botocore==1.34.11 ; python_version >= "3.11" and python_version < "4.0"
bs4==0.0.1 ; python_version >= "3.11" and python_version < "4.0"
certifi==2023.11.17 ; python_version >= "3.11" and python_version < "4.0"
cffi==1.16.0 ; os_name == "nt" and implementation_name != "pypy" and python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
filetype==1.2.0 ; python_version >= "3.11" and python_version < "4.0"
furl==2.1.3 ; python_version >= "3.11" and python_version < "4.0"
greenlet==3.0.3 ; python_version >= "3.11" and python_version < "4.0" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
h11==0.14.0 ; python_version >= "3.11" and python_version < "4.0"
httpcore==1.0.2 ; python_version >= "3.11" and python_version < "4.0"
idna==3.6 ; python_version >= "3.11" and python_version < "4.0"
jmespath==1.0.1 ; python_version >= "3.11" and python_version < "4.0"
mypy-extensions==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
numpy==1.26.2 ; python_version >= "3.12" and python_version < "4.0" or python_version == "3.11"
orderedmultidict==1.0.1 ; python_version >= "3.11" and python_version < "4.0"
outcome==1.3.0.post0 ; python_version >= "3.11" and python_version < "4.0"
packaging==23.2 ; python_version >= "3.11" and python_version < "4.0"
pandas==2.1.4 ; python_version >= "3.11" and python_version < "4.0"
pathspec==0.12.1 ; python_version >= "3.11" and python_version < "4.0"
platformdirs==4.1.0 ; python_version >= "3.11" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.11" and os_name == "nt" and implementation_name != "pypy" and python_version < "4.0"
pydantic-core==2.14.6 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.5.3 ; python_version >= "3.11" and python_version < "4.0"
pysocks==1.7.1 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
pytz==2023.3.post1 ; python_version >= "3.11" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.11" and python_version < "4.0"
retrying==1.3.4 ; python_version >= "3.11" and python_version < "4.0"
s3transfer==0.10.0 ; python_version >= "3.11" and python_version < "4.0"
selenium-stealth==1.0.6 ; python_version >= "3.11" and python_version < "4.0"
selenium==4.16.0 ; python_version >= "3.11" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.11" and python_version < "4.0"
sortedcontainers==2.4.0 ; python_version >= "3.11" and python_version < "4.0"
soupsieve==2.5 ; python_version >= "3.11" and python_version < "4.0"
tqdm==4.66.1 ; python_version >= "3.11" and python_version < "4.0"
trio-websocket==0.11.1 ; python_version >= "3.11" and python_version < "4.0"
trio==0.23.2 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.9.0 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2023.4 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.0.7 ; python_version >= "3.11" and python_version < "4.0"
urllib3[socks]==2.0.7 ; python_version >= "3.11" and python_version < "4.0"
webdriver-manager==4.0.1 ; python_version >= "3.11" and python_version < "4.0"
wsproto==1.2.0 ; python_version >= "3.11" and python_version < "4.0"
xmltodict==0.13.0 ; python_version >= "3.11" and python_version < "4.0"
pymysql
openai
apache-airflow-providers-mongo
httpx[socks]
slack_sdk
pgvector